requirements.txt
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.0
six==1.17.0
tzdata==2025.2
